+++ src/get/getManifest.js
@@ -11,21 +11,25 @@ import { log } from "../log.js";
 export const getManifest = (manifestUrl) => {
   let chunks = undefined;
 
-  return new Promise((resolve, reject) => {
-    get(manifestUrl, (res) => {
+  return new Promise((resolve) => {
+    const req = get(manifestUrl, (res) => {
       res.on("data", (chunk) => {
         chunks += chunk;
       });
 
       res.on("error", (e) => {
         log.error(e);
-        reject(undefined);
+        resolve(undefined);
       });
 
       res.on("end", () => {
         log.debug("Succesfully cached manifest metadata");
         resolve(chunks);
       });
+    });
+    req.on("error", (e) => {
+      log.warn(e);
+      resolve(undefined);
     });
   });
 };